CGIV-8057: courier booking - when creating OrderLabels if we cant find the shipping service name then default to the service code.
We set both the code and name on the label and attempt to look up the name from the code. This wont always work (e.g. Amazon Logistics)
so we should just fall back to using the code (in the case of Amazon Logistics the code is the name anyway)