LIS-22: updated eBay-specific account controller to remove the default Payment Method.
This is a workaround. With a default payment method set when the user first looks at the page all the required fields will appear completed,
if they don't need to change anything then they wont and there'll be nothing to trigger saving the defaults.
Leaving one of the required fields blank will force them to set it and thus trigger the save.
This is a quick fix for us at the expense of being mildly annoying to the user, although only users who wouldnt otherwise need to set anything.