LIS-103: updated create listing Ebay and VariationPicker React components to make the Listing Type selector work with the variation picker.
I had to do all the state changes in one place - VariationPicker - so that the two components weren't fighting it out.
The Listing Type selector itself now just sets listingType in the state which triggers the VariationPicker to recalculate the rest.
Also when different variations are picked that can reset the Listing Type when required.