LIS-92: in the React Select component onClick function replaced hard-coded checks for certain child elements with a data attribute on them.
This means that there's not a blanket affect on those types of nodes, just ones we explicitly want to affect.
Note: the use of data attributes is usually not required in React as you can bind data directly to a callback but as this is potentially a
bubbled event we have no control over the passed data so we had to resort to data attributes in this case.