LIS-92: fix to Select React component to ensure the options and selectedOption are consistently in the right state.
We have to copy them from props to state as the user can manipulate them (both the selected option and the options themselves as they can enter new ones).
This was previously only happening on initialisation and not if the props changed.
Also, when an option was selected we were immediately updating the parent component's state before the select components state had finished updating
which messed things up.