LIS-154: tweak to Products\Listing\MultiCreationService::createListings() to separate out the generation of a unique ID.
Previously the $guid was being passed by referene, set in this method then used again in the calling code.
Passing scalars by reference should generally be avoided as (a) it's confusing to anyone reading the calling code and
(b) the method was effectively returning two values in a roundabout sort of way.