LIS-158: multi-listing creation - improvement to Ebay/CustomItemSpecific to replace local state with dispatching a Redux action.
When a custom item specific was removed we needed to update ReduxForm. Previously we held the input elements in state so that we could call onChange()
on them later. That worked but was hacky and not very Redux-y. Now instead we use Redux's mapDispatchToProps() to create a callback prop that can dispatch a
Redux action to notify ReduxForm of the change.