LIS-192: commercialise listings - improved check for permission to create listings by caching it in the session.
We only do this if the user *does* have permission. This means that when they add the Listings package we don't have to invalidate the cache.
If they remove the package the permission status will still be cached but we've set it to expire after 24 hours.

Also added some logging around this.