LIS-260: product creation - fix to DimensionsTable\Component::isFirstVariationRow() to work consistently.
When it was working was when the user typed in a normal, non-custom attribute field first. This meant that the first property to be assigned to .variations was 'variation-0'.
When the user filled in a custom attribute column first then the first property wast c-table-with-inputs__headings - that screwed up the loop.

The first variation 'ID' (index) will always be zero so we can simply check for that. The reason we didnt do that originally is we thought the user might be able to remove the first (zero'th) row but they can't.