TAC-125: USPS - when creating a ShipStation account from a USPS shipping account we now copy the warehouseId over.
For the other ShipStation carriers we create them an account on ShipEngine with their own API key and warehouse.
For USPS we're not doing that as its all going through CGs account. However, in order to make calls to ShipEngine
we need a ShipEngine account so we create one with our credentials. The warehouses are obviously different per user though
so we're creating them, recording the ID against the USPS account then copying it over to the ShipEngine account on use.