TAC-109: ShipStation - changed the Label\Creator\Factory to pass along the Guzzle Client to the created classes.
This is required because by default we use our Failover Client in place of Guzzle Client but that is set up to work with our internal APIs, not third parties.
We had configured the original Label\Creator to use a normal Guzzle Client but that got lost when we changed to using a factory.