TAC-228: added a commented-out line to the webpack config as a reminder that we'll need it once we're using ES6 modules.
By default Webpack exposes the full `export` object from ES6 modules but we need just the `export.default`.
The config `output.libraryExport = "default"` does this for us.
However, we cant turn it on now becuase if its set while we still have AMD modules as the source then it breaks.