TAC-252: UK Mail - changed the keys of the services in the config to their names rather than their numeric code.
The final tariff code is determined in the mapper anyway so it doesnt matter what we key them by, as long as they're unique.
When they were keyed by their code, because they were numbers, something along the way was reindexing the array and they got
sorted numerically, which brings them out in a non-intuitive order. Changing the keys to their names solves this and has the
bonus of making the config easier to read.