TAC-385: couriers - fixed an edge case where there's only one courier connected and it has dropdown columns which was causing a fatal.
Fetching the options for e.g. Package Type, Insurance Options, etc requires a service. If the user has only one courier we skip the Review
page of the Ship section and send them straight to the Specifics page which means they haven't chosen a service yet. That was causing the
above to blow up. We now handle this case.