TAC-472: fix to PDF generation so that if no custom templates are selected, only the default template, then we don't end up generating an extra blank page.
The default template isn't saved in the DB so doesnt come back in the fetch for templates. If only the default is selected then we still attempt to fetch
the selected Templates and end up with an empty collection. We were still attempting to generate PDFs for them which resulted in a single blank page which
then got merged in with the generated default PDF, leaving a blank page at the end.