MIG-69: fix to Category Template page to correctly fetch Amazon root categories.
These are correctly set as `listable:false` but when fetching them we were filtering to
`listable:true` so they didnt come back.
We didnt see this on Live as the change to make them `listable:false` went out in August
(CGIV-10610) and we've not updated the categories since then but we can see it locally.