TAC-626: Fetch the attachments for messages when populating the data passed to the message centre UI.
This is necessary even if we have the attachments on message entities as embedded data as we might run into caching issues: for instance,
if we now save a message, then save it's attachments internally, when we refetch the message from the internal API, the attachments won't be available as embedded data
because we get the old cached versionthe message entity. Same thing can happen with the thread entity, we wouldn't have the attachments available on the embedded messages
on thread entities.