CGIV-11760 We need to run Collection Request much more often than I thought eariler. Some message from UkMail
The collection request generates a request to the customers collecting depot to send a driver to go make a collection.
This needs to be integrated as a standard requirement for all mutual customers wishing to ship for us.

There are some customers that have a requirement for weekends/bank holidays or out of hours collection and a collection job date will be passed for them.
We refer to this as deferred and it needs authorisation before you can go ahead and configure a customer for this.

Back to the collection call it should be requested prior to creating a label and within the last collection time for the driver.
If the call is made after the LCT then an error will be returned advising of such and to request a collection for the next day.
The collection relates to the customers’ accounts and is specific to that customer, usually a customer will have a collection for all accounts from the same depot.
If a different depot needs to collect you will be advised of this. So doesn’t really matter which account you use as long as that account is the one we are collecting from.